+++ cargo-rbmt/src/bench.rs
@@ -4,7 +4,7 @@
 
 use xshell::Shell;
 
-use crate::environment::{get_workspace_packages, OutputMode, ProgressGuard};
+use crate::environment::{cargo_cmd, get_workspace_packages, OutputMode, ProgressGuard};
 use crate::lock::LockFile;
 use crate::toolchain::{prepare_toolchain, Toolchain};
 
@@ -27,8 +27,7 @@ pub fn run(
         let _dir = sh.push_dir(&package.dir);
 
         // Capture output and show in stdout for verbose mode.
-        let output =
-            rbmt_cmd!(sh, "cargo --locked bench").env("RUSTFLAGS", "--cfg=bench").read()?;
+        let output = cargo_cmd(sh).arg("bench").env("RUSTFLAGS", "--cfg=bench").read()?;
         if matches!(OutputMode::from_env(), OutputMode::Verbose) {
             println!("{}", output);
         }